storybook-testing: unify the stories-* decorator harness
Lifts the generic story substrate into @dxos/storybook-testing —
createStoryDecorators (theme/layout, plugin-manager host with lazy
plugin loading, client init/seeding/snapshot import, runtime-layout
atom, story-args function form) plus shared composable plugins
(makeModuleSurfacesPlugin, StoryAiPlugin, UpdateCompanionStubPlugin)
— and migrates stories-assistant (now a thin domain wrapper keeping
chat/agent creation, skill binding, and the scripted model),
stories-projects, stories-brain, and stories-inbox onto it, deleting
their hand-rolled copies.
