+++ packages/stories/stories-assistant/src/testing/decorators.tsx
@@ -2,20 +2,23 @@
 // Copyright 2025 DXOS.org
 //
 
+//
+// Assistant-domain story harness, layered on `createStoryDecorators` from `@dxos/storybook-testing`
+// (which owns the generic substrate: theme/layout, the plugin manager, client init/seeding/snapshot
+// import, and the runtime-layout atom). This wrapper adds what is assistant-specific: the assistant
+// plugin stack, chat/agent creation, skill binding, and the scripted (offline) model.
+//
+
 import * as Effect from 'effect/Effect';
 import * as Layer from 'effect/Layer';
-import * as Atom from 'effect/unstable/reactivity/Atom';
-import React, { type FC, ReactNode, useEffect, useMemo, useState } from 'react';
+import React, { type ReactNode } from 'react';
 
 import { ScriptedLanguageModel, SERVICES_CONFIG } from '@dxos/ai/testing';
 import * as ActivationEvents from '@dxos/app-framework/ActivationEvents';
 import * as Capabilities from '@dxos/app-framework/Capabilities';
 import * as Capability from '@dxos/app-framework/Capability';
-import * as CapabilityManager from '@dxos/app-framework/CapabilityManager';
 import * as Plugin from '@dxos/app-framework/Plugin';
-import * as PluginManager from '@dxos/app-framework/PluginManager';
-import { type WithPluginManagerOptions, activateDemandGatedModules } from '@dxos/app-framework/testing';
-import { useApp, useCapabilities, useCapability } from '@dxos/app-framework/ui';
+import { useCapabilities, useCapability } from '@dxos/app-framework/ui';
 import * as AppCapabilities from '@dxos/app-toolkit/AppCapabilities';
 import * as AppSpace from '@dxos/app-toolkit/AppSpace';
 import * as GraphPath from '@dxos/app-toolkit/GraphPath';
@@ -33,7 +36,6 @@ import {
   makeDelegationStrategy,
 } from '@dxos/assistant-toolkit';
 import { type Space } from '@dxos/client/echo';
-import { persistentClientServices } from '@dxos/client/testing';
 import * as Instructions from '@dxos/compute/Instructions';
 import * as Operation from '@dxos/compute/Operation';
 import * as OperationHandlerSet from '@dxos/compute/OperationHandlerSet';
@@ -47,36 +49,28 @@ import { EffectEx } from '@dxos/effect';
 import { invariant } from '@dxos/invariant';
 import { DXN } from '@dxos/keys';
 import { AccessToken } from '@dxos/link';
-import { log } from '@dxos/log';
 import * as AssistantOperation from '@dxos/plugin-assistant/AssistantOperation';
 import * as AssistantPlugin from '@dxos/plugin-assistant/AssistantPlugin';
 import { translations as assistantTranslations } from '@dxos/plugin-assistant/translations';
 import * as ClientCapabilities from '@dxos/plugin-client/ClientCapabilities';
 import * as ClientEvents from '@dxos/plugin-client/ClientEvents';
-import * as ClientOptions from '@dxos/plugin-client/ClientOptions';
-import * as ClientPlugin from '@dxos/plugin-client/ClientPlugin';
-import { initializeIdentity } from '@dxos/plugin-client/testing';
 import * as Markdown from '@dxos/plugin-markdown/Markdown';
 import * as MarkdownOperationHandlerSet from '@dxos/plugin-markdown/MarkdownOperationHandlerSet';
 import * as MarkdownSkill from '@dxos/plugin-markdown/MarkdownSkill';
 import { PreviewPlugin } from '@dxos/plugin-preview/testing';
 import * as RoutineCapabilities from '@dxos/plugin-routine/RoutineCapabilities';
 import * as RoutinePlugin from '@dxos/plugin-routine/RoutinePlugin';
-import { corePlugins } from '@dxos/plugin-testing';
-import * as StorybookPlugin from '@dxos/plugin-testing/StorybookPlugin';
 import * as TranscriptionPlugin from '@dxos/plugin-transcription/TranscriptionPlugin';
-import { type Client, Config } from '@dxos/react-client';
+import { Config } from '@dxos/react-client';
 import { useQuery, useSpaces } from '@dxos/react-client/echo';
 import { useAsyncEffect } from '@dxos/react-ui';
 import { translations as debugTranslations } from '@dxos/react-ui-debug/translations';
-import { withLayout, withTheme } from '@dxos/react-ui/testing';
 import { Text } from '@dxos/schema';
-import { type ModuleLayout, StoryLayout } from '@dxos/storybook-testing';
+import { type StoryDecoratorsProps, createStoryDecorators } from '@dxos/storybook-testing';
 import { Outline, Task, TaskSet } from '@dxos/types';
-import { isNonNullable } from '@dxos/util';
+import { Merge, isNonNullable } from '@dxos/util';
 
 import { moduleSurfaces } from '../modules';
-import { initClientFromSpaceSnapshot } from './snapshot';
 
 /** Shared CSF parameters for the assistant story groups (fullscreen canvas + plugin translations). */
 export const storyParameters = {
@@ -108,212 +102,26 @@ export const config = {
   }),
 };
 
-type LazyPluginsResult = {
-  plugins: Plugin.Plugin[];
-  types?: any[];
-};
-
-type DecoratorsProps = {
-  plugins?: Plugin.Plugin[];
-  lazyPlugins?: () => Promise<LazyPluginsResult>;
-  accessTokens?: AccessToken.AccessToken[];
-  /** Import a `.dx.json` space archive instead of creating an empty space. */
-  importSnapshot?: () => Promise<unknown>;
-  onInit?: (props: { client: Client; space: Space }) => Promise<ModuleLayout | void>;
-  /** Skill-definition keys to clone into the space and bind into the latest chat's context. */
-  skills?: string[];
-  /**
-   * Replace the AI service with a scripted (offline, deterministic) model — see
-   * `ScriptedLanguageModel`. Makes the full-stack story runnable without a network AI service,
-   * e.g. from a `play` function in CI; routed scripts can drive cooperating sessions
-   * (supervisor + sub-agents).
-   */
-  scripted?: ScriptedLanguageModel.Script;
-} & (Omit<ClientOptions.ClientPluginOptions, 'onClientInitialized' | 'onSpacesReady'> &
-  Pick<StoryPluginOptions, 'onChatCreated' | 'createAgent'>);
+type DecoratorsProps = Merge<
+  {
+    /** Skill-definition keys to clone into the space and bind into the latest chat's context. */
+    skills?: string[];
+    /**
+     * Replace the AI service with a scripted (offline, deterministic) model — see
+     * `ScriptedLanguageModel`. Makes the full-stack story runnable without a network AI service,
+     * e.g. from a `play` function in CI; routed scripts can drive cooperating sessions
+     * (supervisor + sub-agents).
+     */
+    scripted?: ScriptedLanguageModel.Script;
+  },
+  Omit<StoryDecoratorsProps, 'Wrapper' | 'setupEvents'>,
+  Pick<StoryPluginOptions, 'onChatCreated' | 'createAgent'>
+>;
 
 /**
- * Builds the full plugin list for the plugin manager.
- */
-const buildPluginManagerOptions = ({
-  types = [],
-  plugins = [],
-  accessTokens = [],
-  importSnapshot,
-  onInit,
-  onChatCreated,
-  createAgent,
-  config,
-  scripted,
-  ...props
-}: Omit<DecoratorsProps, 'lazyPlugins'>): WithPluginManagerOptions => {
-  // The `persistent` config preset (see `config.persistent` below) flags itself via
-  // `storage.persistent`; only then do we pay for DEDICATED_WORKER + a dedicated/coordinator
-  // worker pair — every other preset (`remote`/`local`) stays a plain ephemeral client.
-  const clientServices = config?.values.runtime?.client?.storage?.persistent
-    ? persistentClientServices(config)
-    : { config };
-
-  // Shared per-story: `onInit` fills the holder during client-init; the setup module (which holds
-  // the AtomRegistry) copies it into `layoutAtom`, contributed as `StoryLayout.Atom` for the generic
-  // `ModuleContainer` to read.
-  const layoutHolder: { current?: ModuleLayout } = {};
-  const layoutAtom = Atom.make<ModuleLayout | undefined>(undefined);
-
-  return {
-    // SetupSchema registers ECHO schemas so plugin-scoped types are available in stories.
-    // SetupSettings causes plugins (e.g. AssistantPlugin) to contribute settings capabilities
-    // that surfaces like TracePanel read via `useAtomCapability(AssistantCapabilities.Settings)`.
-    plugins: [
-      ...corePlugins(),
-      ClientPlugin.make({
-        types: [
-          AccessToken.AccessToken,
-          Chat.Chat,
-          Collection.Collection,
-          Outline.Outline,
-          Task.Task,
-          TaskSet.TaskSet,
-          Text.Text,
-          Skill.Skill,
-          Operation.PersistentOperation,
-          Markdown.Document,
-          Instructions.Instructions,
-          Trigger.Trigger,
-          ...types,
-        ],
-        onClientInitialized: ({ client }) =>
-          Effect.gen(function* () {
-            log.info('onClientInitialized', { identity: client.halo.identity.get()?.did });
-            // Abort if already initialized.
-            if (client.halo.identity.get()) {
-              return;
-            }
-
-            if (importSnapshot) {
-              yield* initClientFromSpaceSnapshot(importSnapshot)({ client });
-              const space = client.spaces.get()[0];
-              invariant(space, 'No space available after snapshot import.');
-
-              for (const accessToken of accessTokens) {
-                space.db.add(Obj.clone(accessToken));
-              }
-
-              if (onInit) {
-                layoutHolder.current = (yield* Effect.promise(() => onInit({ client, space }))) || undefined;
-              }
-
-              yield* Effect.promise(() => space.db.flush({ indexes: true }));
-              return;
-            }
-
-            const { defaultSpace: space } = yield* initializeIdentity(client);
-
-            // Add tokens.
-            for (const accessToken of accessTokens) {
-              space.db.add(Obj.clone(accessToken));
-            }
-
-            yield* Effect.promise(() => space.db.flush({ indexes: true }));
-            if (onInit) {
-              layoutHolder.current = (yield* Effect.promise(() => onInit({ client, space }))) || undefined;
-            }
-            yield* Effect.promise(() => space.db.flush({ indexes: true }));
-          }),
-        ...clientServices,
-        ...props,
-      }),
-
-      // User plugins.
-      PreviewPlugin.make(),
-      RoutinePlugin.make(),
-      AssistantPlugin.make(
-        scripted ? { aiServiceMiddleware: ScriptedLanguageModel.scriptedAiServiceMiddleware(scripted) } : {},
-      ),
-      TranscriptionPlugin.make(),
-
-      // Test-specific.
-      StorybookPlugin.make({}),
-      StoryPlugin({ onChatCreated, createAgent, layoutAtom, layoutHolder }),
-      ...plugins,
-    ],
-  };
-};
-
-/**
- * Inner component that creates the plugin manager and renders the app.
- * Separated to respect React hooks rules (hooks must be called unconditionally).
- */
-const PluginManagerHost = ({
-  options,
-  children,
-  contextId,
-}: {
-  options: WithPluginManagerOptions;
-  children: ReactNode;
-  contextId: string;
-}) => {
-  const manager = useMemo(() => {
-    const pluginManager = PluginManager.make({
-      pluginLoader: () => Effect.die(new Error('Not implemented')),
-      plugins: options.plugins ?? [],
-      enabled: (options.plugins ?? []).map(({ meta }) => meta.profile.key),
-    });
-
-    // `useApp` contributes these too, but from an effect registered AFTER this component's own —
-    // which kicks off activation — so the startup pass would reach a module requiring `AtomRegistry`
-    // with no provider registered and nothing to wait for. Contribute at construction instead.
-    pluginManager.capabilities.contribute({
-      interface: Capabilities.PluginManager,
-      implementation: pluginManager,
-      module: 'org.dxos.app-framework.plugin-manager',
-    });
-    pluginManager.capabilities.contribute({
-      interface: Capabilities.AtomRegistry,
-      implementation: pluginManager.registry,
-      module: 'org.dxos.app-framework.atom-registry',
-    });
-
-    return pluginManager;
-  }, [options]);
-
-  useEffect(() => {
-    const [capability] = CapabilityManager.expandContributions([
-      Capability.contribute(Capabilities.ReactRoot, {
-        id: contextId,
-        root: () => <>{children}</>,
-      }),
-    ]);
-
-    manager.capabilities.contribute({
-      interface: capability.interface,
-      implementation: capability.implementation,
-      module: 'org.dxos.app-framework.with-plugin-manager.lazy',
-    });
-
-    // A story mounts one surface, so no demand ever reaches the modules gated behind it. The
-    // `withPluginManager` path does this for us; this lazy path builds its own manager and so
-    // must do it too, or Idle-gated contributions (assistant settings) never land and the first
-    // strict `useAtomCapability` read throws.
-    EffectEx.runDetached(activateDemandGatedModules(manager));
-
-    return () => {
-      manager.capabilities.remove(capability.interface, capability.implementation);
-      void EffectEx.runAndForwardErrors(manager.shutdown());
-    };
-  }, [manager, contextId, children]);
-
-  // Forward `setupEvents` (e.g. SetupSettings) so plugins contribute their settings capabilities;
-  // `useApp` is what fires them, and without this the lazy path skips them (the non-lazy
-  // `withPluginManager` path forwards them automatically).
-  const App = useApp({ pluginManager: manager, setupEvents: options.setupEvents });
-  return <App />;
-};
-
-/**
- * Decorator body that binds story-declared skill keys into the latest chat's context. Lives as a
- * decorator (owned by {@link createDecorators}) rather than a container wrapper, so stories render
- * the generic `ModuleContainer` directly.
+ * Decorator body that binds story-declared skill keys into the latest chat's context. Rendered by
+ * the shared harness inside the plugin-manager context (via the `Wrapper` prop), so its capability
+ * hooks always resolve.
  */
 const SkillBinder = ({ skills = [], children }: { skills?: string[]; children: ReactNode }) => {
   const atomRegistry = useCapability(Capabilities.AtomRegistry);
@@ -356,79 +164,49 @@ const SkillBinder = ({ skills = [], children }: { skills?: string[]; children: R
 };
 
 /**
- * Create storybook decorators.
- * Supports lazy plugin loading via the `lazyPlugins` option.
+ * Create storybook decorators for the assistant story groups: the shared harness plus the
+ * assistant plugin stack, chat/agent creation, and skill binding.
  */
-export const createDecorators = ({ config: configProp = config.remote, lazyPlugins, ...props }: DecoratorsProps) => {
-  // Theme + fullscreen layout are common to every story group, so the factory owns them (outermost
-  // decorators) rather than each `meta` repeating a `storyDecorators` array. The plugin manager and
-  // the optional skill binder both render via `PluginManagerHost` (lazy or not) so the binder's
-  // capability hooks always resolve inside the manager context — a trailing decorator would sit
-  // outside it.
-  const host = ((Story: FC, context: { id: string }) => {
-    // Non-lazy stories start with options ready ({ plugins: [] } is truthy); lazy stories wait.
-    const [lazyResult, setLazyResult] = useState<LazyPluginsResult | null>(lazyPlugins ? null : { plugins: [] });
-    useEffect(() => {
-      if (lazyPlugins) {
-        void lazyPlugins().then(setLazyResult);
-      }
-    }, []);
-
-    const options = useMemo(
-      () =>
-        lazyResult
-          ? buildPluginManagerOptions({
-              ...props,
-              config: configProp,
-              plugins: [...(props.plugins ?? []), ...(lazyResult.plugins ?? [])],
-              types: [...(props.types ?? []), ...(lazyResult.types ?? [])],
-            })
-          : null,
-      // `props`/`lazyPlugins` are stable per createDecorators call; the captured config must
-      // still invalidate the memo if a recreated decorator carries a different one.
-      // eslint-disable-next-line react-hooks/exhaustive-deps
-      [lazyResult, configProp],
-    );
-
-    if (!options) {
-      return null;
-    }
-
-    return (
-      <PluginManagerHost options={options} contextId={context.id}>
-        {props.skills?.length ? (
-          <SkillBinder skills={props.skills}>
-            <Story />
-          </SkillBinder>
-        ) : (
-          <Story />
-        )}
-      </PluginManagerHost>
-    );
-  }) as any;
-
-  return [withTheme(), withLayout({ layout: 'fullscreen' }), host];
-};
-
-/**
- * Creates access tokens from environment variables.
- * @param tokens - Record of token sources mapped to their VITE_ prefixed environment variable values
- * @returns Array of AccessToken objects for non-empty token values
- * @example
- * ```tsx
- * const tokens = accessTokensFromEnv({
- *   'exa.ai': process.env.VITE_EXA_API_KEY,
- *   'linear.app': process.env.VITE_LINEAR_API_KEY
- * });
- * ```
- * @note All environment variables should use the VITE_ prefix for proper Vite bundling
- */
-
-export const accessTokensFromEnv = (tokens: Record<string, string | undefined>) => {
-  return Object.entries(tokens)
-    .filter(([, token]) => !!token)
-    .map(([source, token]) => Obj.make(AccessToken.AccessToken, { source, token: token! }));
-};
+export const createDecorators = ({
+  config: configProp = config.remote,
+  skills,
+  scripted,
+  onChatCreated,
+  createAgent,
+  types = [],
+  plugins = [],
+  ...props
+}: DecoratorsProps) =>
+  createStoryDecorators({
+    ...props,
+    config: configProp,
+    types: [
+      AccessToken.AccessToken,
+      Chat.Chat,
+      Collection.Collection,
+      Outline.Outline,
+      Task.Task,
+      TaskSet.TaskSet,
+      Text.Text,
+      Skill.Skill,
+      Operation.PersistentOperation,
+      Markdown.Document,
+      Instructions.Instructions,
+      Trigger.Trigger,
+      ...types,
+    ],
+    plugins: [
+      PreviewPlugin.make(),
+      RoutinePlugin.make(),
+      AssistantPlugin.make(
+        scripted ? { aiServiceMiddleware: ScriptedLanguageModel.scriptedAiServiceMiddleware(scripted) } : {},
+      ),
+      TranscriptionPlugin.make(),
+      StoryPlugin({ onChatCreated, createAgent }),
+      ...plugins,
+    ],
+    Wrapper: skills?.length ? ({ children }) => <SkillBinder skills={skills}>{children}</SkillBinder> : undefined,
+  });
 
 type CreateAgentOptions = {
   name?: string;
@@ -443,10 +221,6 @@ type StoryPluginOptions = {
    * Accepts `true` for defaults, or an options object for name/instructions.
    */
   createAgent?: boolean | CreateAgentOptions;
-
-  /** Shared with `buildPluginManagerOptions` — see the comment where it is created. */
-  layoutAtom: Atom.Writable<ModuleLayout | undefined>;
-  layoutHolder: { current?: ModuleLayout };
 };
 
 const StoryPlugin = Plugin.define<StoryPluginOptions>(
@@ -460,11 +234,6 @@ const StoryPlugin = Plugin.define<StoryPluginOptions>(
     provides: [Capabilities.ReactSurface],
     activate: () => Effect.succeed([Capability.contribute(Capabilities.ReactSurface, moduleSurfaces)]),
   }),
-  Plugin.addModule(({ layoutAtom }) => ({
-    id: 'com.example.plugin.testing.module.layout',
-    provides: [StoryLayout.Atom],
-    activate: () => Effect.succeed([Capability.contribute(StoryLayout.Atom, layoutAtom)]),
-  })),
   Plugin.addModule({
     id: 'com.example.plugin.testing.module.testing',
     // Startup, not the implicit Idle: `AgentServiceSpec` reads `AgentDelegationStrategy` through
@@ -493,7 +262,7 @@ const StoryPlugin = Plugin.define<StoryPluginOptions>(
         ]),
       ]),
   }),
-  Plugin.addModule(({ createAgent, onChatCreated, layoutAtom, layoutHolder }) => ({
+  Plugin.addModule(({ createAgent, onChatCreated }) => ({
     id: 'com.example.plugin.testing.module.setup',
     // Runtime event: the space isn't available until the client observes it.
     activatesOn: ClientEvents.SpacesReady,
@@ -559,12 +328,6 @@ const StoryPlugin = Plugin.define<StoryPluginOptions>(
             Effect.ensuring(Effect.promise(() => binder.close())),
           );
         }
-      }
-
-      // Publish the story layout (built by `onInit`) now that the space + objects exist.
-      if (layoutHolder.current) {
-        const registry = yield* Capabilities.AtomRegistry;
-        registry.set(layoutAtom, layoutHolder.current);
       }
     }),
   })),
