+++ packages/stories/stories-projects/src/testing/decorators.tsx
@@ -3,26 +3,16 @@
 //
 
 import * as Effect from 'effect/Effect';
-import * as Layer from 'effect/Layer';
 
-import { AiService } from '@dxos/ai';
-import { AiServiceTestingPreset } from '@dxos/ai/testing';
-import * as ActivationEvents from '@dxos/app-framework/ActivationEvents';
-import * as Capabilities from '@dxos/app-framework/Capabilities';
-import * as Capability from '@dxos/app-framework/Capability';
-import * as Plugin from '@dxos/app-framework/Plugin';
-import { withPluginManager } from '@dxos/app-framework/testing';
+import type * as Plugin from '@dxos/app-framework/Plugin';
 import * as AppActivationEvents from '@dxos/app-toolkit/AppActivationEvents';
 import * as Instructions from '@dxos/compute/Instructions';
-import * as LayerSpec from '@dxos/compute/LayerSpec';
 import * as Project from '@dxos/compute/Project';
 import * as Routine from '@dxos/compute/Routine';
 import * as Trigger from '@dxos/compute/Trigger';
 import { Collection, Database, Feed, type Type } from '@dxos/echo';
 import { EffectEx } from '@dxos/effect';
 import { mockAiService } from '@dxos/extractor/testing';
-import { DXN } from '@dxos/keys';
-import { ClientPlugin, initializeIdentity } from '@dxos/plugin-client/testing';
 import * as Mailbox from '@dxos/plugin-inbox/Mailbox';
 import { Builder, InboxPlugin } from '@dxos/plugin-inbox/testing';
 import { translations as inboxTranslations } from '@dxos/plugin-inbox/translations';
@@ -31,12 +21,10 @@ import { translations as projectsTranslations } from '@dxos/plugin-projects/tran
 import * as RoutinePlugin from '@dxos/plugin-routine/RoutinePlugin';
 import { translations as routineTranslations } from '@dxos/plugin-routine/translations';
 import { SpacePlugin } from '@dxos/plugin-space/testing';
-import { corePlugins } from '@dxos/plugin-testing';
-import * as StorybookPlugin from '@dxos/plugin-testing/StorybookPlugin';
 import { translations as formTranslations } from '@dxos/react-ui-form/translations';
-import { withLayout, withTheme } from '@dxos/react-ui/testing';
 import { translations as reactUiTranslations } from '@dxos/react-ui/translations';
 import { TagIndex, Text } from '@dxos/schema';
+import { StoryAiPlugin, createStoryDecorators, makeModuleSurfacesPlugin } from '@dxos/storybook-testing';
 import { Message, Organization, Person } from '@dxos/types';
 
 import { moduleSurfaces } from '../modules';
@@ -53,18 +41,6 @@ export const storyParameters = {
     ...reactUiTranslations,
   ],
 };
-
-/** Contributes this package's module surfaces so a story can drive them from a `ModuleContainer` layout. */
-const StoryModulesPlugin = Plugin.define(
-  Plugin.makeMeta({ key: DXN.make('org.dxos.plugin.projects.story.modules'), name: 'Project Story Modules' }),
-).pipe(
-  Plugin.addModule({
-    id: 'project-story-modules',
-    provides: [Capabilities.ReactSurface],
-    activate: () => Effect.succeed([Capability.contribute(Capabilities.ReactSurface, moduleSurfaces)]),
-  }),
-  Plugin.make,
-);
 
 /**
  * Which AI service the story runs against. Absent means none: templates still scaffold and the
@@ -93,101 +69,58 @@ export type DecoratorsProps = {
 };
 
 /**
- * Provides `AiService` on the space-affinity layer the operations resolve at invoke time, matching
- * how the app provisions it — so an operation invoked from a story (fact extraction, a routine's
- * prompt) reaches a model instead of failing to resolve the service.
- */
-const StoryAiPlugin = (kind: StoryAiService) =>
-  Plugin.define(
-    Plugin.makeMeta({ key: DXN.make('org.dxos.plugin.projects.story.ai'), name: 'Project Story AI Service' }),
-  ).pipe(
-    Plugin.addModule({
-      id: 'project-story-ai',
-      // Restart-scoped: the process manager snapshots LayerSpecs once at boot (see AppCapability.layerSpec).
-      activatesOn: ActivationEvents.Startup,
-      provides: [Capabilities.LayerSpec],
-      activate: Capability.makeModule(
-        Effect.fnUntraced(function* () {
-          return [
-            Capability.contribute(
-              Capabilities.LayerSpec,
-              LayerSpec.make({ affinity: 'space', requires: [], provides: [AiService.AiService] }, () =>
-                kind === 'ollama'
-                  ? // `orDie`: a missing/unreachable provider is a story setup fault, not a
-                    // recoverable operation error, and `LayerSpec` requires an empty error channel.
-                    AiServiceTestingPreset('ollama').pipe(Layer.orDie)
-                  : mockAiService({ text: 'A concise summary.' }),
-              ),
-            ),
-          ];
-        }),
-      ),
-    }),
-    Plugin.make,
-  )();
-
-/**
  * Story decorators for a project template scenario: a default space seeded with one mailbox, the
  * plugin set that owns the project machinery, and this package's module surfaces.
  *
  * The plugin list is what populates the skill-definition registry that resolves a skill ref to a
  * display label; without the owning plugin the article's skill rows and picker render empty.
  */
-export const createDecorators = ({ mailboxName, messages, ai, plugins = [], types = [] }: DecoratorsProps) => [
-  withTheme(),
-  withLayout({ layout: 'fullscreen' }),
-  withPluginManager({
+export const createDecorators = ({ mailboxName, messages, ai, plugins = [], types = [] }: DecoratorsProps) =>
+  createStoryDecorators({
     // Skill-definition modules ride the assistant's start event, and nothing here opens assistant UI
     // or materializes a toolkit — without firing it the article's skill rows resolve to blank labels.
     setupEvents: [AppActivationEvents.AssistantStart],
+    types: [
+      Mailbox.Mailbox,
+      Project.Project,
+      Instructions.Instructions,
+      Routine.Routine,
+      Trigger.Trigger,
+      Collection.Collection,
+      // Mailbox and every trigger's feed spec resolve `Feed`; unregistered, the client logs
+      // "Schema not registered" and feed-backed reads (messages, trigger summaries) come up empty.
+      Feed.Feed,
+      TagIndex.TagIndex,
+      Text.Text,
+      // Seeded mail: the messages themselves plus the Person/Organization objects the builder
+      // links them to.
+      Message.Message,
+      Person.Person,
+      Organization.Organization,
+      ...types,
+    ],
+    onInit: async ({ space }) => {
+      const mailbox = space.db.add(Mailbox.make({ name: mailboxName }));
+      if (messages && messages.count > 0) {
+        const feed = await mailbox.feed.load();
+        const built = new Builder()
+          .createMessages(messages.count, { links: { db: space.db }, threads: messages.threads ?? 3 })
+          .build();
+        await EffectEx.runAndForwardErrors(
+          Feed.append(feed, built.messages).pipe(Effect.provide(Database.layer(space.db))),
+        );
+      }
+      await space.db.flush({ indexes: true });
+    },
     plugins: [
-      ...corePlugins(),
-      ClientPlugin.make({
-        types: [
-          Mailbox.Mailbox,
-          Project.Project,
-          Instructions.Instructions,
-          Routine.Routine,
-          Trigger.Trigger,
-          Collection.Collection,
-          // Mailbox and every trigger's feed spec resolve `Feed`; unregistered, the client logs
-          // "Schema not registered" and feed-backed reads (messages, trigger summaries) come up empty.
-          Feed.Feed,
-          TagIndex.TagIndex,
-          Text.Text,
-          // Seeded mail: the messages themselves plus the Person/Organization objects the builder
-          // links them to.
-          Message.Message,
-          Person.Person,
-          Organization.Organization,
-          ...types,
-        ],
-        onClientInitialized: ({ client }) =>
-          Effect.gen(function* () {
-            const { defaultSpace } = yield* initializeIdentity(client);
-            yield* Effect.promise(async () => {
-              const mailbox = defaultSpace.db.add(Mailbox.make({ name: mailboxName }));
-              if (messages && messages.count > 0) {
-                const feed = await mailbox.feed.tryLoad();
-                const built = new Builder()
-                  .createMessages(messages.count, { links: { db: defaultSpace.db }, threads: messages.threads ?? 3 })
-                  .build();
-                await EffectEx.runAndForwardErrors(
-                  Feed.append(feed!, built.messages).pipe(Effect.provide(Database.layer(defaultSpace.db))),
-                );
-              }
-              await defaultSpace.db.flush({ indexes: true });
-            });
-          }),
-      }),
-      StorybookPlugin.make({}),
       SpacePlugin({}),
       InboxPlugin(),
       ProjectsPlugin.make(),
       RoutinePlugin.make(),
-      StoryModulesPlugin(),
-      ...(ai ? [StoryAiPlugin(ai)] : []),
+      makeModuleSurfacesPlugin('org.dxos.plugin.projects.story.modules', moduleSurfaces),
+      ...(ai
+        ? [StoryAiPlugin({ ai: ai === 'ollama' ? 'ollama' : () => mockAiService({ text: 'A concise summary.' }) })]
+        : []),
       ...plugins,
     ],
-  }),
-];
+  });
