+++ packages/stories/storybook-testing/tsconfig.json
@@ -12,6 +12,9 @@
   ],
   "references": [
     {
+      "path": "../../common/effect"
+    },
+    {
       "path": "../../common/invariant"
     },
     {
@@ -21,10 +24,16 @@
       "path": "../../common/log"
     },
     {
+      "path": "../../core/compute/ai"
+    },
+    {
       "path": "../../core/compute/compute"
     },
     {
       "path": "../../core/compute/compute-runtime"
+    },
+    {
+      "path": "../../core/compute/link"
     },
     {
       "path": "../../core/echo/echo"
